+++ CHANGELOG.md
@@ -14,6 +14,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - **Removed** for now removed features.
 - **Fixed** for any bug fixes.
 
+## [11.0.9] - 27-07-2022
+
+## Added
+
+- Added new methods to NumberUtils
+
+## Fixed
+
+- Fixed exception message in NumberRange constructor
+
 ## [11.0.8] - 20-07-2022
 
 ## Added